Move libcore.tlswire to org.conscrypt.tlswire
The TLS wire protocol stuff isn't actually used anywhere except in
Conscrypt, so move it to our package with the plan to delete it from
libcore once this goes into AOSP.

Also refactor two identical sets of methods into a single new
TlsTester class.
