+++ openjdk-integ-tests/src/test/java/org/conscrypt/javax/net/ssl/SSLSocketVersionCompatibilityTest.java
@@ -31,7 +31,6 @@ import static org.junit.Assume.assumeTrue;
 
 import java.io.ByteArrayInputStream;
 import java.io.DataInputStream;
-import java.io.EOFException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -115,15 +114,16 @@ import javax.net.ssl.TrustManager;
 import javax.net.ssl.X509KeyManager;
 import javax.net.ssl.X509TrustManager;
 import libcore.java.security.StandardNames;
-import libcore.tlswire.handshake.ClientHello;
-import libcore.tlswire.handshake.HandshakeMessage;
-import libcore.tlswire.handshake.HelloExtension;
-import libcore.tlswire.handshake.ServerNameHelloExtension;
-import libcore.tlswire.record.TlsProtocols;
-import libcore.tlswire.record.TlsRecord;
 import org.conscrypt.Conscrypt;
 import org.conscrypt.TestUtils;
 import org.conscrypt.java.security.TestKeyStore;
+import org.conscrypt.tlswire.TlsTester;
+import org.conscrypt.tlswire.handshake.ClientHello;
+import org.conscrypt.tlswire.handshake.HandshakeMessage;
+import org.conscrypt.tlswire.handshake.HelloExtension;
+import org.conscrypt.tlswire.handshake.ServerNameHelloExtension;
+import org.conscrypt.tlswire.record.TlsProtocols;
+import org.conscrypt.tlswire.record.TlsRecord;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -1704,7 +1704,7 @@ public class SSLSocketVersionCompatibilityTest {
                 return socket;
             }
         };
-        TlsRecord firstReceivedTlsRecord = captureTlsHandshakeFirstTlsRecord(sslSocketFactory);
+        TlsRecord firstReceivedTlsRecord = TlsTester.captureTlsHandshakeFirstTlsRecord(executor, sslSocketFactory);
         assertEquals("TLS record type", TlsProtocols.HANDSHAKE, firstReceivedTlsRecord.type);
         HandshakeMessage handshakeMessage = HandshakeMessage.read(
                 new DataInputStream(new ByteArrayInputStream(firstReceivedTlsRecord.fragment)));
@@ -1722,8 +1722,8 @@ public class SSLSocketVersionCompatibilityTest {
         ForEachRunner.runNamed(new Callback<SSLSocketFactory>() {
             @Override
             public void run(SSLSocketFactory sslSocketFactory) throws Exception {
-                ClientHello clientHello = SSLSocketVersionCompatibilityTest.this
-                    .captureTlsHandshakeClientHello(sslSocketFactory);
+                ClientHello clientHello = TlsTester
+                    .captureTlsHandshakeClientHello(executor, sslSocketFactory);
                 ServerNameHelloExtension sniExtension =
                     (ServerNameHelloExtension) clientHello.findExtensionByType(
                         HelloExtension.TYPE_SERVER_NAME);
@@ -1749,99 +1749,7 @@ public class SSLSocketVersionCompatibilityTest {
         }
         return result;
     }
-    private ClientHello captureTlsHandshakeClientHello(SSLSocketFactory sslSocketFactory)
-            throws Exception {
-        TlsRecord record = captureTlsHandshakeFirstTlsRecord(sslSocketFactory);
-        assertEquals("TLS record type", TlsProtocols.HANDSHAKE, record.type);
-        ByteArrayInputStream fragmentIn = new ByteArrayInputStream(record.fragment);
-        HandshakeMessage handshakeMessage = HandshakeMessage.read(new DataInputStream(fragmentIn));
-        assertEquals(
-                "HandshakeMessage type", HandshakeMessage.TYPE_CLIENT_HELLO, handshakeMessage.type);
-        // Assert that the fragment does not contain any more messages
-        assertEquals(0, fragmentIn.available());
-        return (ClientHello) handshakeMessage;
-    }
-    private TlsRecord captureTlsHandshakeFirstTlsRecord(SSLSocketFactory sslSocketFactory)
-            throws Exception {
-        byte[] firstReceivedChunk = captureTlsHandshakeFirstTransmittedChunkBytes(sslSocketFactory);
-        ByteArrayInputStream firstReceivedChunkIn = new ByteArrayInputStream(firstReceivedChunk);
-        TlsRecord record = TlsRecord.read(new DataInputStream(firstReceivedChunkIn));
-        // Assert that the chunk does not contain any more data
-        assertEquals(0, firstReceivedChunkIn.available());
-        return record;
-    }
-    @SuppressWarnings("FutureReturnValueIgnored")
-    private byte[] captureTlsHandshakeFirstTransmittedChunkBytes(
-            final SSLSocketFactory sslSocketFactory) throws Exception {
-        // Since there's no straightforward way to obtain a ClientHello from SSLSocket, this test
-        // does the following:
-        // 1. Creates a listening server socket (a plain one rather than a TLS/SSL one).
-        // 2. Creates a client SSLSocket, which connects to the server socket and initiates the
-        //    TLS/SSL handshake.
-        // 3. Makes the server socket accept an incoming connection on the server socket, and reads
-        //    the first chunk of data received. This chunk is assumed to be the ClientHello.
-        // NOTE: Steps 2 and 3 run concurrently.
-        ServerSocket listeningSocket = null;
-        // Some Socket operations are not interruptible via Thread.interrupt for some reason. To
-        // work around, we unblock these sockets using Socket.close.
-        final Socket[] sockets = new Socket[2];
-        try {
-            // 1. Create the listening server socket.
-            listeningSocket = ServerSocketFactory.getDefault().createServerSocket(0);
-            final ServerSocket finalListeningSocket = listeningSocket;
-            // 2. (in background) Wait for an incoming connection and read its first chunk.
-            final Future<byte[]> readFirstReceivedChunkFuture = runAsync(new Callable<byte[]>() {
-                @Override
-                public byte[] call() throws Exception {
-                    Socket socket = finalListeningSocket.accept();
-                    sockets[1] = socket;
-                    try {
-                        byte[] buffer = new byte[64 * 1024];
-                        int bytesRead = socket.getInputStream().read(buffer);
-                        if (bytesRead == -1) {
-                            throw new EOFException("Failed to read anything");
-                        }
-                        return Arrays.copyOf(buffer, bytesRead);
-                    } finally {
-                        closeQuietly(socket);
-                    }
-                }
-            });
-            // 3. Create a client socket, connect it to the server socket, and start the TLS/SSL
-            //    handshake.
-            runAsync(new Callable<Void>() {
-                @Override
-                public Void call() throws Exception {
-                    Socket client = new Socket();
-                    sockets[0] = client;
-                    try {
-                        client.connect(finalListeningSocket.getLocalSocketAddress());
-                        // Initiate the TLS/SSL handshake which is expected to fail as soon as the
-                        // server socket receives a ClientHello.
-                        try {
-                            SSLSocket sslSocket = (SSLSocket) sslSocketFactory.createSocket(client,
-                                    "localhost.localdomain", finalListeningSocket.getLocalPort(),
-                                    true);
-                            sslSocket.startHandshake();
-                            fail();
-                            return null;
-                        } catch (IOException expected) {
-                            // Ignored.
-                        }
-                        return null;
-                    } finally {
-                        closeQuietly(client);
-                    }
-                }
-            });
-            // Wait for the ClientHello to arrive
-            return readFirstReceivedChunkFuture.get(10, TimeUnit.SECONDS);
-        } finally {
-            closeQuietly(listeningSocket);
-            closeQuietly(sockets[0]);
-            closeQuietly(sockets[1]);
-        }
-    }
+
     // http://b/18428603
     @Test
     public void test_SSLSocket_getPortWithSNI() throws Exception {
@@ -2341,26 +2249,6 @@ public class SSLSocketVersionCompatibilityTest {
             return (SSLSession) method.invoke(socket);
         } catch (Exception e) {
             return null;
-        }
-    }
-
-    private static void closeQuietly(Socket socket) {
-        if (socket != null) {
-            try {
-                socket.close();
-            } catch (Exception ignored) {
-                // Ignored.
-            }
-        }
-    }
-
-    private static void closeQuietly(ServerSocket serverSocket) {
-        if (serverSocket != null) {
-            try {
-                serverSocket.close();
-            } catch (Exception ignored) {
-                // Ignored.
-            }
         }
     }
