+++ testing/src/main/java/org/conscrypt/tlswire/handshake/CipherSuite.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package libcore.tlswire.handshake;
+package org.conscrypt.tlswire.handshake;
 import java.util.HashMap;
 import java.util.Map;
 /**
