+++ testing/src/main/java/org/conscrypt/tlswire/handshake/EllipticCurve.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package libcore.tlswire.handshake;
+package org.conscrypt.tlswire.handshake;
 /**
  * {@code EllipticCurve} enum from RFC 4492 section 5.1.1. Curves are assigned
  * via the
