+++ testing/src/main/java/org/conscrypt/tlswire/record/TlsRecord.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package libcore.tlswire.record;
+package org.conscrypt.tlswire.record;
 
 import java.io.DataInput;
 import java.io.IOException;
-import libcore.tlswire.util.TlsProtocolVersion;
+import org.conscrypt.tlswire.util.TlsProtocolVersion;
 
 /**
  * TLS Record Protocol record from TLS 1.2 RFC 5246.
