+++ testing/src/main/java/org/conscrypt/tlswire/util/IoUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package libcore.tlswire.util;
+package org.conscrypt.tlswire.util;
 import java.io.DataInput;
 import java.io.IOException;
 public class IoUtils {
